Reorganized directories to fit django project/app ideology.
