+++ interlecture/interlecture/routing.py
@@ -1,8 +1,6 @@
 from channels.routing import route,include
-import playground.consumers
+import engine.consumers
 
 channel_routing = [
-    route("websocket.receive", playground.consumers.ws_message),
-    route("websocket.connect", playground.consumers.ws_add),
-    route("websocket.disconnect", playground.consumers.ws_disconnect),
+    include(engine.consumers.channel_routing)
   ]
