+++ interlecture/interlecture/settings.py
@@ -56,7 +56,10 @@ ROOT_URLCONF = 'interlecture.urls'
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [os.path.join(BASE_DIR, 'interlecture/templates')],
+        'DIRS': [
+            os.path.join(BASE_DIR, 'interlecture/templates'),
+            os.path.join(BASE_DIR, 'engine/templates')
+          ],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -122,7 +125,7 @@ USE_TZ = True
 STATIC_URL = '/static/'
 
 STATICFILES_DIRS = [
-    os.path.join(BASE_DIR, 'react/static'),
+    os.path.join(BASE_DIR, 'static/'),
 ]
 
 ################################################################
