+++ interlecture/interlecture/templates/base.html
@@ -1,7 +1,8 @@
 <!doctype html>
-<html class="no-js" lang="">
+<html>
   <head>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css">
+    {% block head %}{% endblock %}
   </head>
   <body>
     {% block main %}{% endblock %}
